requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1
nltk==3.8.1
python-docx==1.1.0
PyPDF2==3.0.1
transformers==4.38.2
torch==2.2.1
sentence-transformers==2.5.1 